backport of commit d70c17f7fd2cc3621c571d5213a2b46ffe63c00f
